requirements.txt
python-pptx==1.0.2
streamlit==1.37.1
langchain==0.2.12
langchain-community==0.2.11
PyPDF2==3.0.1
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
langchain-google-genai==1.0.8
networkx==3.3
